refactor folder #JC19-359
