+++ apps/vanilla/src/pages/detail/index.html
@@ -9,8 +9,8 @@
   </head>
   <body class="default-layout">
     <h1 class="title">Detail page</h1>
+    <a href="/">Go to home</a>
     <div class="detail__content">
-        <a href="/">Go to home</a>
     </div>
     <script type="module" src="./detail.ts"></script>
   </body>
