+++ apps/vanilla/src/services/authService.ts
@@ -1,17 +1,12 @@
 import { LoginMapper } from '@js-camp/core/mappers/login.mapper';
 import { RegisterMapper } from '@js-camp/core/mappers/register.mapper';
-import { UserMapper } from '@js-camp/core/mappers/user.mapper';
 import { Login } from '@js-camp/core/models/login';
 import { Register } from '@js-camp/core/models/register';
-import { User } from '@js-camp/core/models/user';
 
 import { api } from '../api/API';
-import { Token } from '../scripts/constants';
 import { Helpers } from '../scripts/helpers';
 
 import { ErrorService } from './errorService';
-
-import { StorageService } from './storageService';
 
 export namespace AuthService {
 
@@ -38,13 +33,6 @@ export namespace AuthService {
   /** Logout service.*/
   export function logout(): void {
     Helpers.clearToken();
-  }
-
-  /** Get user.*/
-  export async function getUser(): Promise<User> {
-    // eslint-disable-next-line @typescript-eslint/naming-convention
-    const { data } = await api.get('/users/profile/', { headers: { Authorization: `Bearer ${StorageService.get(Token.Access)}` } });
-    return UserMapper.fromDto(data);
   }
 
   /**
